rename: DTO, in/out 폴더 분리
