+++ src/app.controller.ts
@@ -1,9 +1,9 @@
 import { Body, Controller, Get, Post } from '@nestjs/common';
 import { ApiBody, ApiOperation, ApiResponse } from '@nestjs/swagger';
 
-import { CreateUserDto } from '@src/user/dto/create-user.dto';
+import { CreateUserDto } from '@src/user/dto/in/create-user.dto';
 import { UserService } from '@src/user/user.service';
-import { PostUserDto } from '@src/user/dto/post-user.dto';
+import { PostUserDto } from '@src/user/dto/in/post-user.dto';
 
 import { AppService } from '@src/app.service';
