+++ src/user/dto/in/update-user.dto.ts
@@ -1,4 +1,4 @@
 import { PartialType } from '@nestjs/mapped-types';
-import { CreateUserDto } from '@src/user/dto/create-user.dto';
+import { CreateUserDto } from '@src/user/dto/in/create-user.dto';
 
 export class UpdateUserDto extends PartialType(CreateUserDto) {}
