+++ src/user/user.service.ts
@@ -4,8 +4,8 @@ import { EntityRepository } from '@mikro-orm/mysql';
 import { InjectRepository } from '@mikro-orm/nestjs';
 import { HttpException, Injectable } from '@nestjs/common';
 
-import { CreateUserDto } from '@src/user/dto/create-user.dto';
-import { UpdateUserDto } from '@src/user/dto/update-user.dto';
+import { CreateUserDto } from '@src/user/dto/in/create-user.dto';
+import { UpdateUserDto } from '@src/user/dto/in/update-user.dto';
 import { User } from '@src/user/entities/user.entity';
 import { ConfigService } from '@nestjs/config';
 import { ServerConfig } from '@src/common/config';
